Add SABs

+++ src/hotspot/share/gc/shared/c2/barrierSetC2.hpp
@@ -105,7 +105,7 @@ protected:
   Node*             _base;
   C2AccessValuePtr& _addr;
   Node*             _raw_access;
-  uint8_t           _barrier_data;
+  uint16_t          _barrier_data;
 
   void fixup_decorators();
 
@@ -129,9 +129,9 @@ public:
   bool is_raw() const             { return (_decorators & AS_RAW) != 0; }
   Node* raw_access() const        { return _raw_access; }
 
-  uint8_t barrier_data() const        { return _barrier_data; }
-  void set_barrier_data(uint8_t data) { _barrier_data = data; }
-  void add_barrier_data(uint8_t data) { _barrier_data |= data; }
+  uint16_t barrier_data() const        { return _barrier_data; }
+  void set_barrier_data(uint16_t data) { _barrier_data = data; }
+  void add_barrier_data(uint16_t data) { _barrier_data |= data; }
 
   void set_raw_access(Node* raw_access) { _raw_access = raw_access; }
   virtual void set_memory() {} // no-op for normal accesses, but not for atomic accesses.
